Limit effort manually and make it possible to disable the safety controller

Closes #1017

See merge request project-march/march!1023

+++ ros1/src/hardware/interface/march_hardware/src/motor_controller/motor_controller.cpp
@@ -157,7 +157,7 @@ void MotorController::actuate(float target)
     }
 }
 
-float MotorController::effortMultiplicationConstant()
+double MotorController::effortMultiplicationConstant()
 {
     return 1.0;
 }
